[WIP] Add options.accountIndex to Seed.get_key
